+++ tests/testthat/test-connection.R
@@ -9,6 +9,7 @@ test_that("can connect to DataSpace", {
 
 if ("DataSpaceConnection" %in% class(con)) {
   con_names <- c(".__enclos_env__",
+                 "availableGroups",
                  "availableStudies",
                  "config",
                  "clone",
@@ -22,9 +23,9 @@ if ("DataSpaceConnection" %in% class(con)) {
   if (identical(names(con), con_names)) {
     test_that("`print`", {
       cap_output <- capture.output(con$print())
-      expect_length(cap_output, 8)
+      expect_length(cap_output, 9)
 
-      if (length(cap_output) == 8) {
+      if (length(cap_output) == 9) {
         expect_equal(cap_output[1], "<DataSpaceConnection>")
         expect_equal(cap_output[2], "  URL: https://dataspace.cavd.org")
         expect_match(cap_output[3], "  User: \\S+@\\S+")
@@ -33,6 +34,7 @@ if ("DataSpaceConnection" %in% class(con)) {
         expect_match(cap_output[6], "  - \\d+ subjects")
         expect_match(cap_output[7], "  - \\d+ assays")
         expect_match(cap_output[8], "  - \\d+ data points")
+        expect_match(cap_output[9], "  Available groups: \\d+")
       }
     })
 
@@ -75,6 +77,10 @@ if ("DataSpaceConnection" %in% class(con)) {
 
       expect_is(cavd, "DataSpaceStudy")
       expect_is(cavd, "R6")
+
+      expect_error(con$getStudy("cvd0"))
+      expect_error(con$getStudy("", 0))
+      expect_error(con$getStudy("cvd208", 208))
     })
   }
 }